adds contributing.md, renames test folder, moves index under src dir
