+++ package.json
@@ -5,7 +5,7 @@
   "main": "index.js",
   "scripts": {
     "start": "node index.js",
-    "test": "jscs ./ && istanbul cover --report lcov --dir coverage/ _mocha ./test --grep ./test/*.js -- --colors",
+    "test": "jscs ./ && istanbul cover --report lcov --dir coverage/ _mocha ./tests --grep ./tests/*.js -- --colors",
     "lint": "jscs ./"
   },
   "repository": {
